+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 - notification are now over a modal
 - multi select dates for event copy is now possible
+- hide unofficial packages from sponsor period edit/create form
 
 ## v0.7.16 - 2024-09-06
 ### Changed